- Applied Action Controlers

+++ LINQ_Review/Model/YearSet.cs
@@ -9,12 +9,12 @@ namespace LINQ_Review.Model
     struct YearSet
     {
         public int Year { get; private set; }
-        public int CapitalExpendituresPriceIndicator { get; private set; }
-        public int ConstructionAssemblyWorksIndicator { get; private set; }
-        public int InvestnebtPurchasesIndicator { get; private set; }
-        public int OtherExpendituresIndicator { get; private set; }
+        public double CapitalExpendituresPriceIndicator { get; private set; }
+        public double ConstructionAssemblyWorksIndicator { get; private set; }
+        public double InvestnebtPurchasesIndicator { get; private set; }
+        public double OtherExpendituresIndicator { get; private set; }
 
-        public YearSet(int year, int CEP, int CAW, int IP, int OE)
+        public YearSet(int year, double CEP, double CAW, double IP, double OE)
         {
             Year = year;
             CapitalExpendituresPriceIndicator = CEP;
